fmap

map foieGras fitted or predicted locations, with or without Argos observations, optionally apply a different projection

fmap(
  x,
  y = NULL,
  what = c("fitted", "predicted"),
  conf = TRUE,
  obs = FALSE,
  obs.shp = 17,
  by.date = TRUE,
  crs = NULL,
  ext.rng = c(0.05, 0.05),
  size = 0.25,
  col = "black",
  lines = FALSE,
  landfill = grey(0.6),
  pal = "Zissou1",
  rev = FALSE
)

Arguments

x
a foieGras ssm fit object with class fG_ssm

y
optionally, a foieGras mpm fit object with class fG_mpm; default is NULL

what
specify which location estimates to map: fitted or predicted

conf
include confidence regions around estimated location (logical; default = TRUE, unless y is an mpm fit object then conf is FALSE)

obs
include Argos observations on map (logical; default = FALSE)

obs.shp
point shape for observations (default = 17)

by.date
when mapping single tracks, should locations be coloured by date (logical; default = TRUE if nrow(x) == 1 else FALSE)

crs
proj4string for re-projecting locations, if NULL the default projection ("+proj=merc") for the fitting the SSM will be used

ext.rng
factors to extend the plot range in x and y dimensions (can exceed 1)

size
size of estimated location points (size = NA will draw no points). Optionally, a vector of length 2 with size of observed locations given by 2nd value (ignored if obs = FALSE)

col
colour of observed locations (ignored if obs = FALSE)

lines
logical indicating if lines are added to connect estimated locations (default = FALSE)

landfill
colour to use for land (default = grey(0.6))

pal
hcl.colors palette to use (default: "Zissou1"; type hcl.pals() for options)

rev
reverse colour palette (logical)