Coerce input data into expected foieGras format

format data by mapping supplied variable names to those expected by fit_ssm(), and ensuring variables are put into the expected order. Can be run manually by user as a data pre-processing step prior to calling fit_ssm() or can be called automatically by fit_ssm(). In the latter case, any custom variable names must be declared as arguments to fit_ssm(); see examples, below.

format_data(
  x,
  id = "id",
  date = "date",
  lc = "lc",
  coord = c("lon", "lat"),
  epar = c("smaj", "smin", "eor"),
  sderr = c("lonerr", "laterr"),
  tz = "UTC"
)

Arguments

x

input data

id

the name (as a quoted character string) of id variable: a unique identifier for individual (animal) track data sets.

date

the name (as a quoted character string)of the date/time variable: date and time (as YYYY-MM-DD HH:MM:SS) of each observation.

lc

the name (as a quoted character string) of the location quality class variable: Argos location quality class (values in the set: 3,2,1,0,"A","B","Z"). Can also include "G" for GPS data and/or "GL" for light-level geolocation (GLS) and other data types.

coord

the names (as quoted character strings) of the location coordinate variables: defaults are c("lon","lat"), but could also be c("x","y") for planar coordinates; or if input data is an sf object then "geometry". If input data is an sf object then coord is set to "geometry" by default.

epar

the names (as quoted character strings) of the Argos error ellipse parameters: defaults are "smaj" (ellipse semi-major axis), "smin" (ellipse semi-minor axis), and "eor" (ellipse orientation). Ignored if these variables are missing from the input data.

sderr

the names (as quoted character strings) of provided standard errors in longitude and latitude: defaults are "lonerr", "laterr". Typically, these are only provided for processed light-level geolocation data. Ignored if these variables are missing from the input data.

tz

the timezone the applies to the data/time variable if they are not in tz = 'UTC'. A list of valid timezone names can be viewed via OlsonNames()

Value

a data.frame or sf-tibble of input data in expected foieGras format. Additional columns required by fit_ssm(), if missing, will be added to the formatted tibble: smaj, smin, eor, lonerr, and laterr.

Examples

## as a data pre-processing step
data(sese2_n)
head(sese2_n, 5)
#> # A tibble: 5 × 5
#>   longitude latitude time                lc    id       
#>       <dbl>    <dbl> <chr>               <fct> <chr>    
#> 1      72.5    -50.2 2009-02-01 17:50:46 A     ct36-E-09
#> 2      73.0    -50.4 2009-02-02 03:30:26 A     ct36-E-09
#> 3      73.8    -50.8 2009-02-02 17:50:48 B     ct36-E-09
#> 4      74.6    -51.2 2009-02-03 07:39:08 A     ct36-E-09
#> 5      74.9    -51.4 2009-02-03 15:29:59 A     ct36-E-09
d <- format_data(sese2_n, date = "time", coord = c("longitude","latitude"), 
tz = "America/Halifax")
fit <- fit_ssm(d, model = "crw", time.step = 24)
#> fitting crw SSM to 2 tracks...
#>  pars:   0 0 0 0       pars:   -0.67346 -0.33112 -0.65867 0.05447       pars:   -1.27106 0.26973 -0.15797 0.23098       pars:   -2.0383 -0.33471 -0.20486 0.44022       pars:   -1.41541 -0.11659 -0.42043 0.24543       pars:   -1.86518 0.06573 -0.41249 0.30523       pars:   -1.54018 0.0184 -0.38073 0.26286       pars:   -1.66333 -0.10252 -0.30478 0.27348       pars:   -1.5675 -0.06755 -0.3882 0.2642       pars:   -1.63936 -0.04265 -0.34003 0.25483       pars:   -1.5675 -0.06755 -0.3882 0.2642      
#>  pars:   0 0 0 0       pars:   -0.81269 0.03777 0.57011 0.11438       pars:   -2.93073 0.13621 2.05593 0.41248       pars:   -4.2749 0.78502 0.67878 0.78341       pars:   -2.17768 0.33238 0.61295 0.37815       pars:   -3.1248 -0.33284 0.94127 1.13583       pars:   -2.31371 0.27966 0.67512 0.45068       pars:   -2.44851 0.32961 0.6582 0.54791       pars:   -2.46729 0.32518 0.74439 0.69826       pars:   -2.48723 0.24956 0.65496 0.82591       pars:   -2.48964 0.32528 0.68069 0.715       pars:   -2.48843 0.29391 0.69913 0.77427       pars:   -2.48964 0.32528 0.68069 0.715      

## called automatically within fit_ssm()
fit <- fit_ssm(sese2_n, date = "time", coord = c("longitude", "latitude"), 
tz = "America/Halifax", model = "crw", time.step = 24)
#> fitting crw SSM to 2 tracks...
#>  pars:   0 0 0 0       pars:   -0.67346 -0.33112 -0.65867 0.05447       pars:   -1.27106 0.26973 -0.15797 0.23098       pars:   -2.0383 -0.33471 -0.20486 0.44022       pars:   -1.41541 -0.11659 -0.42043 0.24543       pars:   -1.86518 0.06573 -0.41249 0.30523       pars:   -1.54018 0.0184 -0.38073 0.26286       pars:   -1.66333 -0.10252 -0.30478 0.27348       pars:   -1.5675 -0.06755 -0.3882 0.2642       pars:   -1.63936 -0.04265 -0.34003 0.25483       pars:   -1.5675 -0.06755 -0.3882 0.2642      
#>  pars:   0 0 0 0       pars:   -0.81269 0.03777 0.57011 0.11438       pars:   -2.93073 0.13621 2.05593 0.41248       pars:   -4.2749 0.78502 0.67878 0.78341       pars:   -2.17768 0.33238 0.61295 0.37815       pars:   -3.1248 -0.33284 0.94127 1.13583       pars:   -2.31371 0.27966 0.67512 0.45068       pars:   -2.44851 0.32961 0.6582 0.54791       pars:   -2.46729 0.32518 0.74439 0.69826       pars:   -2.48723 0.24956 0.65496 0.82591       pars:   -2.48964 0.32528 0.68069 0.715       pars:   -2.48843 0.29391 0.69913 0.77427       pars:   -2.48964 0.32528 0.68069 0.715