join an mpm-estimated behavioural index to ssm-predicted locations

join() joins ssm-predicted locations and mpm-estimated behavioural index into a single tibble. If the ssm-predicted tibble is a projected sf object then the output of join will also be an sf object (default). This can be avoided by using as_sf = FALSE.

join(
  ssm,
  mpm,
  what.ssm = "predicted",
  as_sf = FALSE,
  normalise = FALSE,
  group = FALSE
)

Arguments

ssm

a foieGras ssm fitted model object

mpm

a foieGras mpm fitted model object

what.ssm

specifies whether ssm predicted or fitted values are to be extracted

as_sf

logical; if FALSE then return a tibble with un-projected lonlat coordinates, otherwise return an sf tibble

normalise

logical; if output includes a move persistence estimate, should g (the move persistence index) be normalised to have minimum = 0 and maximum = 1 (default = FALSE).

group

logical; should g be normalised among individuals as a group, a 'relative g', or separately to highlight regions of lowest and highest move persistence along a track (default = FALSE).

Value

a single tbl with all individuals

Examples

## load example foieGras fit objects (to save time)
## generate a ssm fit object
xs <- fit_ssm(sese2, spdf=FALSE, model = "rw", time.step=72)
#> fitting rw...
#>  pars:   3.49751 3.31944 -1.07766 0 0 0       pars:   2.84962 2.56683 -1.18419 -0.03286 -0.03728 -0.00252       pars:   2.11333 1.71153 -1.30525 -0.07021 -0.07965 -0.00538       pars:   2.31445 1.5811 -0.95617 1.6249 0.82896 0.82295       pars:   2.43486 1.93501 -1.12217 0.64037 0.30848 0.33687       pars:   1.58897 1.77348 -1.14293 0.80415 0.11788 0.72974       pars:   2.1475 1.32516 -1.27399 0.59089 -0.18334 1.27123       pars:   1.92651 1.70326 -1.1691 0.71947 0.08867 0.77963       pars:   1.89882 1.93515 -1.28014 0.5603 0.07336 0.9736       pars:   1.96862 1.79059 -1.30064 0.62275 0.0589 1.02069       pars:   1.90661 1.75178 -1.37144 0.65507 0.06743 1.16572       pars:   1.82299 1.69944 -1.46692 0.69867 0.07894 1.36129       pars:   1.99784 1.76538 -1.64705 0.56898 0.08895 1.66959       pars:   2.00066 1.91015 -1.80686 0.78331 -0.04286 1.93326       pars:   1.92856 1.81708 -1.68239 0.65433 0.07432 1.75359       pars:   1.91374 1.73757 -1.68573 0.61723 0.12038 1.86858       pars:   1.97046 1.79738 -1.71012 0.65186 0.06828 1.978       pars:   1.91419 1.77467 -1.73231 0.6159 0.08889 2.10985       pars:   1.97776 1.7356 -1.78782 0.64872 0.16835 2.19491       pars:   1.97048 1.7485 -1.74145 0.6387 0.10054 2.14072       pars:   1.9451 1.78324 -1.75941 0.63989 0.14586 2.17702       pars:   1.94015 1.76018 -1.77019 0.62616 0.09979 2.22764       pars:   1.9427 1.78664 -1.78202 0.63391 0.12018 2.29263       pars:   1.9427 1.78664 -1.78202 0.63391 0.12018 2.29263      
#>  pars:   3.45621 4.3327 -1.58431 0 0 0       pars:   2.77999 3.62992 -1.8051 -0.00858 -0.001 -0.00092       pars:   0.75132 1.52158 -2.46746 -0.03431 -0.004 -0.00367       pars:   -0.49836 2.9153 -1.66699 3.24332 0.78362 0.69767       pars:   1.03625 2.61726 -2.01555 0.94735 0.23087 0.20549       pars:   2.04631 1.54736 -1.51597 1.20395 0.40277 0.40197       pars:   1.2246 2.03871 -2.0029 0.75829 0.21509 0.203       pars:   0.84354 1.6331 -2.01299 0.4487 0.19819 0.22306       pars:   1.06128 1.86487 -2.00722 0.6256 0.20785 0.2116       pars:   0.86433 2.00382 -1.90753 0.58153 0.24963 0.26564       pars:   1.0119 1.89971 -1.98223 0.61455 0.21832 0.22515       pars:   1.00959 1.84882 -1.94617 0.60393 0.23071 0.24819       pars:   0.98305 1.88041 -1.90309 0.59867 0.25033 0.27505       pars:   1.00032 1.83975 -1.86473 0.58882 0.26699 0.30495       pars:   0.97964 1.86566 -1.82285 0.57659 0.28922 0.33954       pars:   0.99374 1.82878 -1.78947 0.56983 0.30795 0.37982       pars:   0.97609 1.8496 -1.75526 0.55973 0.32937 0.42689       pars:   0.98946 1.82126 -1.73235 0.54929 0.34306 0.48064       pars:   0.97379 1.83635 -1.71001 0.54716 0.35536 0.54033       pars:   0.98915 1.81964 -1.69729 0.52673 0.36061 0.60008       pars:   0.96707 1.82022 -1.68034 0.5544 0.37294 0.65484       pars:   0.98693 1.81556 -1.65866 0.51122 0.38288 0.69777       pars:   0.96671 1.84156 -1.63556 0.51851 0.39202 0.75195       pars:   0.98693 1.81556 -1.65866 0.51122 0.38288 0.69777      
xm <- fit_mpm(xs, what = "p", model = "jmpm")
#> fitting jmpm...
#>  pars:   0 0 0       pars:   -0.63038 -0.77416 0.05745       pars:   -0.21961 -0.90381 0.95993       pars:   0.1024 -0.34798 0.19353       pars:   -0.12912 -0.74761 0.74455       pars:   -0.17853 -0.8329 0.86216       pars:   -0.23363 -0.801 0.86597       pars:   -0.27635 -0.61987 0.76146       pars:   -0.40458 -0.79384 0.62587       pars:   -0.36412 -0.6719 0.40545       pars:   -0.38583 -0.70377 0.60569       pars:   -0.43697 -0.67572 0.577       pars:   -0.38583 -0.70377 0.60569      

## join predicted values as an un-projected tibble
xsm <- join(xs, xm)
xsm
#> # A tibble: 58 × 11
#>    id        date                  lon   lat      x      y  x.se  y.se logit_g
#>    <chr>     <dttm>              <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl>   <dbl>
#>  1 ct36-E-09 2009-02-01 17:00:00  72.6 -50.2  8079. -6446.  34.6  20.0   4.10 
#>  2 ct36-E-09 2009-02-04 17:00:00  76.3 -52.1  8492. -6783.  27.2  16.1   4.10 
#>  3 ct36-E-09 2009-02-07 17:00:00  81.8 -54.4  9106. -7209.  41.0  25.6   4.10 
#>  4 ct36-E-09 2009-02-10 17:00:00  86.1 -56.7  9589. -7656.  32.0  18.7   3.12 
#>  5 ct36-E-09 2009-02-13 17:00:00  91.0 -59.1 10130. -8166.  31.9  18.4   3.56 
#>  6 ct36-E-09 2009-02-16 17:00:00  95.5 -61.5 10636. -8705.  25.5  16.1   3.27 
#>  7 ct36-E-09 2009-02-19 17:00:00 101.  -64.2 11228. -9352.  29.9  17.7   3.08 
#>  8 ct36-E-09 2009-02-22 17:00:00 106.  -65.1 11746. -9583.  24.2  18.9   1.03 
#>  9 ct36-E-09 2009-02-25 17:00:00 109.  -65.7 12154. -9747.  17.1  13.7   0.988
#> 10 ct36-E-09 2009-02-28 17:00:00 110.  -66.2 12261. -9893.  23.6  15.5  -0.806
#> # … with 48 more rows, and 2 more variables: logit_g.se <dbl>, g <dbl>