map_multi_track_mp

map foieGras-estimated locations from multiple tracks with a behavioural index

map_multi_track_mp(
  map_type,
  obs_sf,
  conf_sf,
  line_sf,
  loc_sf,
  extents,
  buffer,
  aes,
  silent,
  ...
)

Arguments

map_type

coastline data to be used in maps

obs_sf

observed location geometry

conf_sf

confidence ellipses around estimated locations, if specified

line_sf

track line, if specified

loc_sf

estimated location geometry

extents

map extents

buffer

distance (in km) to buffer locations for subsetting land polygons (default = 10000). If map extents are expanded by many factors then the buffer distance may need to be increased, otherwise this should not be used.

aes

a list of map aesthetics (size, shape, col, fill, alpha) to be applied, in order, to: 1) estimated locations,; 2) confidence ellipses; 3) track lines; 4) observed locations; 5) land regions; 6) water regions

silent

logical; map silently (default = FALSE)

...

additional arguments passed to ggspatial::annotation_map_tile