fit the move persistence model to regularized location data

generates initial values for model parameters and unobserved gamma's; structures data and initial values for C++ TMB template; fits move persistence model; minimizes the joint log-likelihood via the selected optimizer (nlminb or optim); structures and passes output object to fit_mpm

mpmf(
  x,
  model = c("jmpm", "mpm"),
  control = mpm_control(),
  inner.control = NULL
)

Arguments

x

temporally regularized location data, eg. output from fit_ssm

model

specify whether MPM is to be fit with unpooled ("mpm") or pooled ("jmpm") RW variance(s).

control

list of control settings for the outer optimizer (see mpm_control for details)

inner.control

list of control settings for the inner optimization (see ?TMB::MakeADFUN for additional details)

verbose

is deprecated, use ssm_control(verbose = 1) instead, see ssm_control for details

optim

is deprecated, use ssm_control(optim = "optim") instead, see ssm_control for details

optMeth

is deprecated, use ssm_control(method = "L-BFGS-B") instead, see ssm_control for details

Details

called by fit_mpm, not intended for general use. see ?fit_mpm.