Re-route a movement path around land using the pathroutr package

A wrapper function that uses the pathroutr package https://jmlondon.github.io/pathroutr/ to re-route movement paths that cross a land barrier. The current implementation will take either the output from a fit_ssm model or the simulations generated by simfit.

route_path(
  x,
  what = c("fitted", "predicted"),
  map_scale = 50,
  dist = 50000,
  append = TRUE,
  ...
)

Arguments

x

either a ssm fit object or a simfit object containing simulated paths

what

if using a ssm object should the fitted (typically irregular in time) or predicted (typically regular in time) locations be re-routed.

map_scale

scale of rnaturalearth map to use for land mass: one of 110, 50 (default), or 10. Note that map_scale = 10 is only available if you have the rnaturalearthhires package installed see: https://github.com/ropensci/rnaturalearthhires

dist

buffer distance (m) to add around track locations. The convex hull of these buffered locations defines the size of land polygon used to aid re-routing of points on land. Larger buffers can result in longer computation times. See London (2020) for further details. The default buffer distance is a constant 50000 m.

append

should re-routed locations be appended to the ssm (ssm fit) object (default = TRUE), or returned as a tibble.

...

additional arguments passed to pathroutr::prt_visgraph

Details

route_path uses rnaturalearth::ne_countries at the medium (50) scale, by default, to generate a land barrier. For efficient computation, route_path clips the polygons to the buffered bounds (set by dist (in m) of the movement track(s).

When the input is a ssm object route_path can append the re-routed path locations to the ssm (ssm fit) object. This is useful when move persistence is to be estimated from the re-routed locations via fit_mpm, or tracks are to be visualised with map. route_path can also return a standalone tibble of the re-routed path with the same number of locations as either the original fitted or predicted locations.

When the re-routed path is appended to the ssm object, the path can be extracted using the grab function, e.g. grab(fit, what = "rerouted").

When the input is a simfit object then route_path returns the same object but with the locations within each simulation re-routed.

We recommend that users working on complex rerouting problems and/or requiring higher resolution land barrier data work with the pathroutr package directly by first exctracting foieGras-estimated locations with grab. Higher resolution land barrier data (polygon shapefiles) must be obtained independently.

References

Josh M. London. (2020) pathroutr: An R Package for (Re-)Routing Paths Around Barriers (Version v0.2.1) https://doi.org/10.5281/zenodo.4321827

Examples

fit <- fit_ssm(sese1, vmax = 4, model = "crw", time.step = 24)
#> fitting crw...
#>  pars:   0 0 0 0       pars:   -0.08444 -0.6562 -0.74977 0.01049       pars:   0.60691 -0.30123 -1.21402 0.43536       pars:   0.13025 -0.54597 -0.89394 0.14242       pars:   0.26113 -0.78913 -0.79861 0.2477       pars:   0.46583 -0.63521 -0.89295 0.3958       pars:   0.53372 -0.75758 -0.89446 0.673       pars:   0.61565 -0.70933 -0.9104 0.5513       pars:   0.61565 -0.70933 -0.9104 0.5513      
fit <- route_path(fit, what = "predicted")
grab(fit, what = "rerouted")
#> # A tibble: 58 × 8
#>    id        date                  lon   lat     x      y      x.se      y.se
#>    <chr>     <dttm>              <dbl> <dbl> <dbl>  <dbl>     <dbl>     <dbl>
#>  1 ct36-E-09 2009-02-01 17:00:00  72.5 -50.2 8068. -6450. 0.0000100 0.0000100
#>  2 ct36-E-09 2009-02-02 17:00:00  73.8 -50.8 8217. -6544. 5.67      4.74     
#>  3 ct36-E-09 2009-02-03 17:00:00  75.0 -51.4 8349. -6659. 6.25      4.64     
#>  4 ct36-E-09 2009-02-04 17:00:00  76.2 -52.1 8486. -6781. 8.02      5.92     
#>  5 ct36-E-09 2009-02-05 17:00:00  78.0 -52.7 8678. -6900. 7.54      6.09     
#>  6 ct36-E-09 2009-02-06 17:00:00  80.1 -53.6 8919. -7053. 5.40      4.32     
#>  7 ct36-E-09 2009-02-07 17:00:00  81.9 -54.3 9118. -7200. 8.66      6.41     
#>  8 ct36-E-09 2009-02-08 17:00:00  83.6 -55.1 9303. -7349. 6.69      5.20     
#>  9 ct36-E-09 2009-02-09 17:00:00  85.0 -55.8 9462. -7481. 7.74      6.04     
#> 10 ct36-E-09 2009-02-10 17:00:00  86.2 -56.7 9596. -7665. 6.97      5.35     
#> # … with 48 more rows

trs <- simfit(fit, what = "predicted", reps = 5)
rrt_sims <- route_path(trs)