Add similarity flag value to a simfit object

Calculate the similarity between a single path generated by simfit and the observed path used in the SSM fit. Similarity flag values are based on two interpretations of the difference in distance and bearing between the two paths.

similarity_flag(track, sim_track, flag = 2)

Arguments

track

a dataframe containing longitude and latitudes of the observed path used in the SSM fit

sim_track

a dataframe containing the longitude and latitude of a single simulated path from simfit

flag

set to either 1 or 2 to define similarity index (see details)

Value

a single value representing the similarity in distance and bearing between the two paths

Details

flag = 1 will use an index based on Hazen (2017)
flag = 2 (the default) will use a custom index