Control Values for fit_ssm.

ssm_control selects the numerical minimizer, method, associated control parameters, and parameter bounds used by fit_ssm.

ssm_control(
  optim = c("nlminb", "optim"),
  method = c("L-BFGS-B", "BFGS", "Nelder-Mead", "CG", "SANN", "Brent"),
  lower = NULL,
  upper = NULL,
  verbose = 1,
  se = FALSE,
  ...
)

Arguments

optim
the numerical optimizer used in the fit

method
if optim = "optim" then the optimization method to be used can be one of "BFGS", "L-BFGS-B", "Nelder-Mead", "CG", "SANN", or "Brent" see optim for details

lower
a list named parameter lower bounds, if NULL then built in defaults are used when method = "L-BFGS-B". Possible parameter names are: l_sigma a vector of length 2, log scale; l_rho_p a scalar, logit scale; l_D a scalar, log scale; l_psi a scalar, log scale; l_tau a vector of length 2, log scale; l_rho_o a scalar, logit scale

upper
a list of named parameter upper bounds, if NULL then built in defaults are used when method = "L-BFGS-B". Possible parameter names are same as lower

verbose
integer; report progress during minimization: 0 = silent; 1 = optimizer trace; 2 = parameter trace (default))

se
logical; should standard errors for speed estimates be calculated (default = FALSE). Turning this on will slow down computation time but provide SE's for speed-along-track calculations

...
control parameters for the chosen optimizer

Value

Returns a list with components

optim

the name of the numerical optimizer as a string, "nlminb" or "optim"

method

optimization method to be used

lower

named list of lower parameter bounds

upper

named list of upper parameter bounds

verbose

level of tracing information to be reported

control

list of control parameters for the optimizer

Details

The optimizer used to minimize the objective function is selected by the optim argument. Additional control parameters specific to the chosen optimizer are specified via the dots argument. See nlminb and optim for available options. Adapted from S. Wotherspoon https://github.com/SWotherspoon/RWalc/blob/master/R/RWalc.R

See also

Examples

fit <- fit_ssm(ellie,
vmax = 4,
model = "crw",
time.step = 72,
control = ssm_control(
    optim = "nlminb",
    eval.max = 2000)
    )
#> fitting crw...
#>  pars:   0 0       pars:   -0.99985 -0.01712       pars:   -3.99941 -0.06849       pars:   -2.23053 -0.94822       pars:   -3.08498 -0.08792       pars:   -3.23225 -0.99063       pars:   -3.39923 -3.28097       pars:   -3.35868 -3.70886       pars:   -3.35868 -3.70886