grab tibble's by name from a foieGras model object

grab() lets you obtain fitted, predicted, rerouted or data tibble's from a compound tibble created when fitting to multiple individual data sets. The specified tibble's are appended to a single output tibble.

grab(x, what = "fitted", as_sf = FALSE)

Arguments

x

a foieGras ssm or mpm model object

what

the tibble to be grabbed; either fitted, predicted, rerouted (ssm only), or data (single letters can be used).

as_sf

logical; if FALSE (default) then return a tibble with un-projected lonlat coordinates, otherwise return an sf tibble. Ignored if x is an mpm model object.

Value

a tibble with all individual tibble's appended

Examples

## generate an ssm fit object
xs <- fit_ssm(sese2, spdf=FALSE, model = "rw", time.step=72, 
control = ssm_control(verbose = 0))
#> 
#> 

## grab predicted values as an un-projected tibble
preds <- grab(xs, what = "predicted")