join an mpm-estimated behavioural index to ssm-predicted locations

join() joins ssm-predicted locations and mpm-estimated behavioural index into a single tibble. If the ssm-predicted tibble is a projected sf object then the output of join will also be an sf object (default). This can be avoided by using as_sf = FALSE.

join(ssm, mpm, what.ssm = "predicted", as_sf = TRUE)

Arguments

ssm

a foieGras ssm fitted model object

mpm

a foieGras mpm fitted model object

what.ssm

specifies whether ssm predicted or fitted values are to be extracted

as_sf

logical; if FALSE then return a tibble with un-projected lonlat coordinates, otherwise return an sf tibble

Value

a single tbl with all individuals

Examples

## load example foieGras fit objects (to save time)
## generate a fG_ssm fit object
xs <- fit_ssm(sese2, spdf=FALSE, model = "rw", time.step=72, 
control = ssm_control(se = FALSE, verbose = 0))
#> 
#> 
data(xm)

## join predicted values as an un-projected tibble
xsm <- join(xs, xm, as_sf = FALSE)
xsm
#> # A tibble: 58 × 10
#>    id      date                  lon   lat      x      y  x.se  y.se     g  g.se
#>    <chr>   <dttm>              <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 ct36-E… 2009-02-01 17:00:00  72.6 -50.2  8079. -6446.  34.6  20.0 0.983 3.28 
#>  2 ct36-E… 2009-02-04 17:00:00  76.3 -52.1  8492. -6783.  27.2  16.1 0.983 2.77 
#>  3 ct36-E… 2009-02-07 17:00:00  81.8 -54.4  9106. -7209.  41.0  25.6 0.983 2.14 
#>  4 ct36-E… 2009-02-10 17:00:00  86.1 -56.7  9589. -7656.  32.0  18.7 0.956 3.49 
#>  5 ct36-E… 2009-02-13 17:00:00  91.0 -59.1 10130. -8166.  31.9  18.4 0.971 1.88 
#>  6 ct36-E… 2009-02-16 17:00:00  95.5 -61.5 10636. -8705.  25.5  16.1 0.962 1.58 
#>  7 ct36-E… 2009-02-19 17:00:00 101.  -64.2 11228. -9352.  29.9  17.7 0.956 0.972
#>  8 ct36-E… 2009-02-22 17:00:00 106.  -65.1 11746. -9583.  24.2  18.9 0.748 0.564
#>  9 ct36-E… 2009-02-25 17:00:00 109.  -65.7 12154. -9747.  17.1  13.7 0.729 0.589
#> 10 ct36-E… 2009-02-28 17:00:00 110.  -66.2 12261. -9893.  23.6  15.5 0.309 0.709
#> # … with 48 more rows