Re-route a movement path around land using the pathroutr package

A wrapper function that uses the pathroutr package https://jmlondon.github.io/pathroutr/ to re-route movement paths that cross a land barrier. The current implementation will take either the output from a fit_ssm model or the simulations generated by sim_fit.

route_path(
  x,
  what = c("fitted", "predicted"),
  map_scale = 50,
  dist = 50000,
  append = TRUE,
  ...
)

Arguments

x

either a ssm fit object or a sim_fit object containing simulated paths

what

if using a ssm object should the fitted (typically irregular in time) or predicted (typically regular in time) locations be re-routed.

map_scale

scale of rnaturalearth map to use for land mass: one of 110, 50 (default), or 10. Note that map_scale = 10 is only available if you have the rnaturalearthhires package installed see: https://github.com/ropensci/rnaturalearthhires

dist

buffer distance (m) to add around track locations. The convex hull of these buffered locations defines the size of land polygon used to aid re-routing of points on land. Larger buffers can result in longer computation times. See London (2020) for further details. The default buffer distance is a constant 50000 m.

append

should re-routed locations be appended to the ssm (ssm fit) object (default = TRUE), or returned as a tibble.

...

additional arguments passed to pathroutr::prt_visgraph

Details

route_path uses rnaturalearth::ne_countries at the medium (50) scale, by default, to generate a land barrier. For efficient computation, route_path clips the polygons to the buffered bounds (set by dist (in m) of the movement track(s).

When the input is a ssm object route_path can append the re-routed path locations to the ssm (ssm fit) object. This is useful when move persistence is to be estimated from the re-routed locations via fit_mpm, or tracks are to be visualised with map. route_path can also return a standalone tibble of the re-routed path with the same number of locations as either the original fitted or predicted locations.

When the re-routed path is appended to the ssm object, the path can be extracted using the grab function, e.g. grab(fit, what = "rerouted").

When the input is a sim_fit object then route_path returns the same object but with the locations within each simulation re-routed.

We recommend that users working on complex rerouting problems and/or requiring higher resolution land barrier data work with the pathroutr package directly by first exctracting foieGras-estimated locations with grab. Higher resolution land barrier data (polygon shapefiles) must be obtained independently.

References

Josh M. London. (2020) pathroutr: An R Package for (Re-)Routing Paths Around Barriers (Version v0.2.1) https://zenodo.org/record/5522909#.YnPxEy_b1qs

Examples

# if 'pathroutr' is installed then ok to use route_path()
if(requireNamespace("pathroutr", quietly = TRUE)) {
  fit <- fit_ssm(ellie, vmax = 4, model = "crw", time.step = 24)
  fit <- route_path(fit, what = "predicted")
  grab(fit, what = "rerouted")
}
#> fitting crw SSM to 1 tracks...
#>  pars:   0 0       pars:   -0.99985 -0.01735       pars:   -3.9994 -0.0694       pars:   -1.96165 -0.9409       pars:   -2.90468 -0.08719       pars:   -3.06264 -1.08284       pars:   -3.2409 -3.52513       pars:   -3.18326 -3.61981       pars:   -3.18326 -3.61981      
#> # A tibble: 108 × 8
#>    id    date                  lon   lat      x      y       x.se       y.se
#>    <chr> <dttm>              <dbl> <dbl>  <dbl>  <dbl>      <dbl>      <dbl>
#>  1 54591 2012-03-05 05:00:00  111. -66.4 12309. -9956.  0.0000100  0.0000100
#>  2 54591 2012-03-06 05:00:00  110. -66.4 12300. -9947. 18.3        0.0103   
#>  3 54591 2012-03-07 05:00:00  110. -66.4 12297. -9956. 26.4       18.2      
#>  4 54591 2012-03-08 05:00:00  110. -66.5 12297. -9970. 22.1       17.5      
#>  5 54591 2012-03-09 05:00:00  110. -66.5 12299. -9972.  8.23       1.23     
#>  6 54591 2012-03-10 05:00:00  110. -66.4 12300. -9955.  5.21       2.88     
#>  7 54591 2012-03-11 05:00:00  110. -66.4 12291. -9958. 18.7       13.6      
#>  8 54591 2012-03-12 05:00:00  110. -66.5 12284. -9962. 18.5        4.46     
#>  9 54591 2012-03-13 05:00:00  110. -66.4 12287. -9952.  6.08       4.62     
#> 10 54591 2012-03-14 05:00:00  110. -66.4 12291. -9939.  4.52       4.13     
#> # … with 98 more rows